+++ pyproject.toml
@@ -110,7 +110,10 @@ pytest-xdist = "^3"
 ruff = "^0"
 sh = "^2"
 "shandy-sqlfmt" = {version = "^0", extras=["shandy-sqlfmt[jinjafmt]"]}
+types-PyYaml = "^6"
 types-tabulate = "^0"
+types-toml = "^0"
+ipykernel = "^6.29.5"
 
 [tool.poetry.group.docs.dependencies]
 mike = "^2"
@@ -154,7 +157,7 @@ exclude = [
     "dist",
     "node_modules",
     "site-packages",
-    "src/dbt_bouncer/dbt_cloud",
+    "src/dbt_bouncer/artifact_parsers/dbt_cloud",
     "venv",
 ]
 line-length = 88
