+++ src/dbt_bouncer/checks/catalog/check_catalog_sources.py
@@ -7,7 +7,7 @@ from dbt_bouncer.check_base import BaseCheck
 if TYPE_CHECKING:
     import warnings
 
-    from dbt_bouncer.parsers import (
+    from dbt_bouncer.artifact_parsers.parsers_common import (
         DbtBouncerSourceBase,
     )
