+++ src/dbt_bouncer/checks/catalog/check_columns.py
@@ -9,7 +9,10 @@ if TYPE_CHECKING:
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=UserWarning)
         from dbt_artifacts_parser.parsers.catalog.catalog_v1 import CatalogTable
-    from dbt_bouncer.parsers import DbtBouncerModelBase, DbtBouncerTestBase
+    from dbt_bouncer.artifact_parsers.parsers_common import (
+        DbtBouncerModelBase,
+        DbtBouncerTestBase,
+    )
 
 from pydantic import Field
