+++ src/dbt_bouncer/checks/manifest/check_semantic_models.py
@@ -8,7 +8,7 @@ from dbt_bouncer.check_base import BaseCheck
 if TYPE_CHECKING:
     import warnings
 
-    from dbt_bouncer.parsers import (
+    from dbt_bouncer.artifact_parsers.parsers_common import (
         DbtBouncerModelBase,
         DbtBouncerSemanticModelBase,
     )
