+++ src/dbt_bouncer/checks/manifest/check_unit_tests.py
@@ -11,7 +11,7 @@ from dbt_bouncer.check_base import BaseCheck
 if TYPE_CHECKING:
     import warnings
 
-    from dbt_bouncer.parsers import DbtBouncerManifest
+    from dbt_bouncer.artifact_parsers.parsers_common import DbtBouncerManifest
 
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=UserWarning)
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
 from dbt_bouncer.utils import object_in_path
 
 if TYPE_CHECKING:
-    from dbt_bouncer.parsers import (
+    from dbt_bouncer.artifact_parsers.parsers_common import (
         DbtBouncerManifest,
         DbtBouncerModelBase,
     )
