+++ tests/unit/checks/catalog/test_catalog_sources.py
@@ -8,10 +8,12 @@ with warnings.catch_warnings():
     from dbt_artifacts_parser.parsers.catalog.catalog_v1 import CatalogTable
     from dbt_artifacts_parser.parsers.manifest.manifest_v12 import Sources
 
+from dbt_bouncer.artifact_parsers.parsers_manifest import (
+    DbtBouncerSourceBase,  # noqa: F401
+)
 from dbt_bouncer.checks.catalog.check_catalog_sources import (
     CheckSourceColumnsAreAllDocumented,
 )
-from dbt_bouncer.parsers import DbtBouncerSourceBase  # noqa: F401
 
 CheckSourceColumnsAreAllDocumented.model_rebuild()
