+++ tests/unit/checks/manifest/test_exposures.py
@@ -7,14 +7,14 @@ with warnings.catch_warnings():
     warnings.filterwarnings("ignore", category=UserWarning)
     from dbt_artifacts_parser.parsers.manifest.manifest_v12 import Exposures, Nodes4
 
-from dbt_bouncer.checks.manifest.check_exposures import (
-    CheckExposureOnNonPublicModels,
-    CheckExposureOnView,
-)
-from dbt_bouncer.parsers import (  # noqa: F401
+from dbt_bouncer.artifact_parsers.parsers_manifest import (  # noqa: F401
     DbtBouncerExposureBase,
     DbtBouncerModel,
     DbtBouncerModelBase,
+)
+from dbt_bouncer.checks.manifest.check_exposures import (
+    CheckExposureOnNonPublicModels,
+    CheckExposureOnView,
 )
 
 CheckExposureOnNonPublicModels.model_rebuild()
