+++ tests/unit/checks/manifest/test_lineage.py
@@ -7,12 +7,15 @@ with warnings.catch_warnings():
     warnings.filterwarnings("ignore", category=UserWarning)
     from dbt_artifacts_parser.parsers.manifest.manifest_v12 import Nodes4
 
+from dbt_bouncer.artifact_parsers.parsers_manifest import (  # noqa: F401
+    DbtBouncerManifest,
+    DbtBouncerModelBase,
+)
 from dbt_bouncer.checks.manifest.check_lineage import (
     CheckLineagePermittedUpstreamModels,
     CheckLineageSeedCannotBeUsed,
     CheckLineageSourceCannotBeUsed,
 )
-from dbt_bouncer.parsers import DbtBouncerManifest, DbtBouncerModelBase  # noqa: F401
 
 CheckLineagePermittedUpstreamModels.model_rebuild()
 CheckLineageSeedCannotBeUsed.model_rebuild()
