+++ tests/unit/checks/manifest/test_unit_tests.py
@@ -7,16 +7,20 @@ with warnings.catch_warnings():
     warnings.filterwarnings("ignore", category=UserWarning)
     from dbt_artifacts_parser.parsers.manifest.manifest_v12 import UnitTests
 
-    from dbt_bouncer.parsers import DbtBouncerModelBase  # noqa: F401
+    from dbt_bouncer.artifact_parsers.parsers_manifest import (
+        DbtBouncerModelBase,  # noqa: F401
+    )
 
 import warnings
 
+from dbt_bouncer.artifact_parsers.parsers_manifest import (
+    DbtBouncerManifest,  # noqa: F401
+)
 from dbt_bouncer.checks.manifest.check_unit_tests import (
     CheckUnitTestCoverage,
     CheckUnitTestExpectFormats,
     CheckUnitTestGivenFormats,
 )
-from dbt_bouncer.parsers import DbtBouncerManifest  # noqa: F401
 
 with warnings.catch_warnings():
     warnings.filterwarnings("ignore", category=UserWarning)
